% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateATE.R
\name{estimateATE}
\alias{estimateATE}
\title{estimateATE function}
\usage{
estimateATE(
  dataset,
  bexp,
  exp.status = 1,
  cexp,
  fmethod.replace = TRUE,
  distbuf = 0.1,
  exp.included = TRUE,
  long,
  lat,
  formulaPS,
  PSmethod = "mgcv.GAM.strata",
  formulaCGPS,
  smethod = "caliper",
  caliper_bw = 0.1,
  smethod.replace = FALSE,
  formulaDisease,
  family,
  bs.N,
  bs.replace = TRUE,
  varilist,
  corrmethod = "Pearson"
)
}
\arguments{
\item{dataset}{a dataset object.}

\item{bexp}{a character string indicating the name of the binary exposure. Use apostrophe like "VariableName"}

\item{exp.status}{a numeric vector indicating the value indicating exposed units. Defalut=1}

\item{cexp}{a character string indicating the name of the continuous exposure. Use apostrophe like "VariableName"}

\item{fmethod.replace}{an indicator of whether matching by distance is done with replacement. Default is TRUE. If FALSE, matching is done without replacement. If FALSE, note that the output of this function may differ by the order of observation units in the original dataset.}

\item{distbuf}{a numeric vector indicating the buffer distance by which exposed units and unexposed units are matched}

\item{exp.included}{an indicator of whether exposed units are matched with not only unexposed units but also other exposed units. Defalut is TRUE. If FALSE, exposed units are matched with only unexposed units. See details}

\item{long}{a character string indicating the name of the longitude variable of observation units}

\item{lat}{a character string indicating the name of the latitude variable of observation units}

\item{formulaPS}{a character string indicating the formula of propensity score estimation. "the binary exposure variable ~ variableA+variableB+variableC+s(long,lat)"}

\item{PSmethod}{a character string indicating which method is used to estimate propensity score. Default is "mgcv.GAM.strata". Options also include "mgcv.GAM", "gnm.GNM". If "mgcv.GAM.strata" is used, a generalized additive model in mgcv package is used. Dummy variables indicating the stratum of matched exposed and unexposed units is included in the model. If "mgcv.GAM" is used, a generalized additive model is used but the dummy variables are not included. If "gnm.GNM" is used, the dummy variables are considered using a generalized non-linear model in gnm package, which is computationally more efficient. However, this does not allow for smoothing terms that can be used in generalized additive models in mgcv package.}

\item{formulaCGPS}{a character string indicating the formula of generalized propensity score estimation}

\item{smethod}{method a character string indicating the matching method used to conduct matching by GPS. Default is "caliper". Options include "nearest" (nearest neighbor matching) and "caliper" (caliper matching)}

\item{caliper_bw}{a numeric vector indicating caliper bandwidth. Default is 0.1. If method is "nearest", this parameter is ignored.}

\item{smethod.replace}{an indicator of whether matching by GPS is done with replacement. Default is TRUE. If FALSE, matching is done without replacement. If FALSE, note that the output of this function may differ by the order of observation units in the original dataset.}

\item{formulaDisease}{a character string indicating the formula of the disease model.}

\item{family}{a character string indicating the error distribution and link function to be used in the disease model.}

\item{bs.N}{a numeric vector indicating the number of bootstrapping samples}

\item{bs.replace}{a character string indicating whether bootstrapping is done with replacement. Default is TRUE.}

\item{varilist}{a character vector indicating variable names for which you wish to compute standardized mean difference. List variable names as a vector like c("VariableA","VariableB")}

\item{corrmethod}{a character string indicating which correlation coefficient is to be computed. These include "Pearson" (default), "Spearman", "Polychoric", or "Polyserial". For tetrachoric use "Polychoric" and for biserial use "Polyserial". This relies on wCorr::weightedCorr}
}
\description{
This function matches exposed units and unexposed units by a pre-specified buffer distance (Euclidean distance)
}
\examples{
estimateATE()
}
