% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.nearest.R
\name{match.nearest}
\alias{match.nearest}
\title{match.nearest function}
\usage{
match.nearest(data, bexp, gpsname, index.status = 1, replace2 = TRUE)
}
\description{
This internal function matches units by the generalized propensity score computed by cgpsmatch function. This function is used only within cgpsmatch function.
}
\examples{
match.nearest()
}
