% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.nearestcaliper.R
\name{match.nearestcaliper}
\alias{match.nearestcaliper}
\title{match.nearestcaliper function}
\usage{
match.nearestcaliper(
  data,
  bexp,
  gpsname,
  caliper_bw,
  exp.status = 1,
  index.status = 1,
  replace2 = TRUE
)
}
\description{
This internal function matches units by the generalized propensity score computed by cgpsmatch function. This function is used only within cgpsmatch function.
}
\examples{
match.caliper()
}
