% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchdist.R
\name{matchdist}
\alias{matchdist}
\title{matchdist function}
\usage{
matchdist(
  data,
  bexp,
  x,
  y,
  exp.status = 1,
  distbuf = 0.1,
  exp.included = TRUE,
  replace = TRUE
)
}
\arguments{
\item{data}{a dataset object.}

\item{bexp}{a character string indicating the name of the binary exposure. Use apostrophe like "VariableName"}

\item{x}{a numeric vector indicating longitude of observation units}

\item{y}{a numeric vector indicating latitude of observation units}

\item{exp.status}{a numeric vector indicating the value indicating exposed units. Defalut=1}

\item{distbuf}{a numeric vector indicating the buffer distance by which exposed units and unexposed units are matched}

\item{exp.included}{an indicator of whether exposed units are matched with not only unexposed units but also other exposed units. Defalut is TRUE. If FALSE, exposed units are matched with only unexposed units. See details}

\item{replace}{an indicator of whether matching is done with replacement. Default is TRUE. If FALSE, matching is done without replacement. If FALSE, note that the output of this function may differ by the order of observation units in the original dataset.}
}
\description{
This function matches exposed units and unexposed units by a pre-specified buffer distance (Euclidian distance)
}
\examples{
matchdist()
}
