% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd}
\alias{smd}
\title{smd function}
\usage{
smd(data, bexp, exp.status = 1, varinames = NULL, weightname = NULL)
}
\arguments{
\item{data}{a dataset object.}

\item{bexp}{a character string indicating the name of binary exposure. Use apostrophe like "VariableName"}

\item{exp.status}{a numeric vector indicating the value indicating exposed units. Defalut=1}

\item{varinames}{a character vector indicating variable names for which you wish to compute standardized mean difference. List variable names as a vector like c("VariableA","VariableB")}

\item{weightname}{the name of the weight variable. This weight is calculated by cgps.match. See cgps.match function. If NULL, weight is not given (i.e., all observations are equally weighted)}
}
\description{
This function calcualtes standardized mean difference in variable(s) in a matched dataset.
}
\examples{
smd(test.dat,"exposure",exp.status=1,varinames=c("X1","X2","X3","U"))
smd()
}
