% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.sd.R
\name{w.sd}
\alias{w.sd}
\title{w.sd. function}
\usage{
w.sd(x, w)
}
\arguments{
\item{x}{is a variable to be calculated; w is a weight.
w.sd()}
}
\description{
This internal function calculates weighted standard deviation
}
