% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.model.cv.R
\name{xgb.model.cv}
\alias{xgb.model.cv}
\title{xgb.model.cv function}
\usage{
xgb.model.cv(
  data,
  cv.objective,
  cv.nround = 1000,
  cv.nfold = 10,
  cv.max_depth = 5,
  cv.eta = 0.1,
  cv.nthread = 1,
  cv.subsample = 0.5,
  cv.gamma = 0.1,
  cv.eval_metric,
  cv.colsample_bytree = 1,
  cv.min_child_weight = 1,
  early_stopping_rounds = 10,
  cv.local.N = 100
)
}
\description{
This internal function explores optimal parameters in xgb.train using cross-validation
xgb.model.cv()
}
