% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.model.fit.cv.R
\name{xgb.model.fit.cv}
\alias{xgb.model.fit.cv}
\title{xgb.model.fit.cv function}
\usage{
xgb.model.fit.cv(
  data,
  cv.objective,
  cv.nround = 10000,
  cv.nfold = 5,
  cv.max_depth = c(4, 5),
  cv.eta = c(0.01, 0.05, 0.1),
  cv.nthread = 1,
  cv.subsample = 1,
  cv.eval_metric,
  cv.colsample_bytree = 1,
  cv.min_child_weight = 1,
  early_stopping_rounds = 50,
  cv.local.N = 10
)
}
\description{
This internal function explores optimal parameters in xgb.train using cross-validation
xgb.model.fit.cv()
}
