% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateATT.R
\name{estimateATT}
\alias{estimateATT}
\title{estimateATT function}
\usage{
estimateATT(
  dataset,
  bexp,
  exp.status = 1,
  cexp,
  fmethod.replace = TRUE,
  distbuf = 0.1,
  exp.included = TRUE,
  long,
  lat,
  PS.method = "mgcv.GAM",
  PS.formula,
  PS.max_depth = 5,
  PS.eta = 0.1,
  PS.verbose = 0,
  PS.nthread = 3,
  PS.eval_metric = "auc",
  PS.objective = "binary:logistic",
  PS.nrounds = 50,
  CGPS.method = "mgcv.GAM",
  CGPS.formula,
  CGPS.max_depth = 5,
  CGPS.eta = 0.1,
  CGPS.verbose = 0,
  CGPS.nthread = 3,
  CGPS.eval_metric = "rmse",
  CGPS.objective = "reg:squarederror",
  CGPS.nrounds = 50,
  smethod = "caliper",
  caliper_bw = 0.1,
  smethod.replace = FALSE,
  weight.cutoff = 10,
  formulaDisease,
  family,
  bs.N,
  bs.replace = TRUE,
  varilist,
  corrmethod = "Pearson"
)
}
\arguments{
\item{dataset}{a dataset object.}

\item{bexp}{a character string indicating the name of the binary exposure. Use apostrophe like "VariableName"}

\item{exp.status}{a numeric vector indicating the value indicating exposed units. Defalut=1}

\item{cexp}{a character string indicating the name of the continuous exposure. Use apostrophe like "VariableName"}

\item{fmethod.replace}{an indicator of whether matching by distance is done with replacement. Default is TRUE. If FALSE, matching is done without replacement. If FALSE, note that the output of this function may differ by the order of observation units in the original dataset.}

\item{distbuf}{a numeric vector indicating the buffer distance by which exposed units and unexposed units are matched}

\item{exp.included}{an indicator of whether exposed units are matched with not only unexposed units but also other exposed units. Defalut is TRUE. If FALSE, exposed units are matched with only unexposed units. See details}

\item{long}{a character string indicating the name of the longitude variable of observation units}

\item{lat}{a character string indicating the name of the latitude variable of observation units}

\item{PS.method}{a character string or a vector of variable names, indicating the method of propensity score estimation. Options include "mgcv.GAM" and "xgboost". Default is "mgcv.GAM"}

\item{PS.formula}{a character string indicating the formula of propensity score estimation. For PSmethod="mgcv.GAM", this string must be like "the binary exposure variable ~ variableA+variableB+variableC+s(long,lat)". For PSmethod="xgboost", this must be a vector like c("variableA", "variableB", "variableC", "long", "lat")}

\item{PS.max_depth}{(xgboost only) a numeric vector indicating maximum depth of a tree. Default is 5}

\item{PS.eta}{(xgboost only) a numeric vector indicating the learning rate. Default is 0.1}

\item{PS.verbose}{(xgboost only) a numeric vector indicating whether xgboost will stay silent. Default is 0 (silent). If 1, it will print information about performance}

\item{PS.nthread}{(xgboost only) a numeric vector indicating the number of thread. See xgb.train}

\item{PS.eval_metric}{(xgboost only) a character string indicating evaluation metrics for validation data. Default is "auc". See xgb.train}

\item{PS.objective}{(xgboost only) a character string indicating the objective function. Default is "binary:logistic". See xgb.train}

\item{PS.nrounds}{(xgboost only) a numeric vector indicating the number of rounds. Default is 50. See xgb.train.}

\item{CGPS.method}{a character string or a vector of variable names, indicating the method of conditional propensity score estimation. Options include "mgcv.GAM" and "xgboost". Default is "mgcv.GAM"}

\item{CGPS.formula}{a character string indicating the formula of conditional generalized propensity score estimation}

\item{CGPS.max_depth}{(xgboost only) a numeric vector indicating maximum depth of a tree. Default is 5}

\item{CGPS.eta}{(xgboost only) a numeric vector indicating the learning rate. Default is 0.1}

\item{CGPS.verbose}{(xgboost only) a numeric vector indicating whether xgboost will stay silent. Default is 0 (silent). If 1, it will print information about performance}

\item{CGPS.nthread}{(xgboost only) a numeric vector indicating the number of thread. See xgb.train}

\item{CGPS.eval_metric}{(xgboost only) a character string indicating evaluation metrics for validation data. Default is "rmse". See xgb.train}

\item{CGPS.objective}{(xgboost only) a character string indicating the objective function. Default is "reg:squarederror". See xgb.train}

\item{CGPS.nrounds}{(xgboost only) a numeric vector indicating the number of rounds. Default is 50. See xgb.train.}

\item{smethod}{method a character string indicating the matching method used to conduct matching by GPS. Default is "caliper". Options include "nearest" (nearest neighbor matching) and "caliper" (caliper matching)}

\item{caliper_bw}{a numeric vector indicating caliper bandwidth. Default is 0.1. If method is "nearest", this parameter is ignored.}

\item{smethod.replace}{an indicator of whether matching by GPS is done with replacement. Default is TRUE. If FALSE, matching is done without replacement. If FALSE, note that the output of this function may differ by the order of observation units in the original dataset.}

\item{weight.cutoff}{a numeric vector indicating the cutoff value of the weight. Default is 10.}

\item{formulaDisease}{a character string indicating the formula of the disease model.}

\item{family}{a character string indicating the error distribution and link function to be used in the disease model.}

\item{bs.N}{a numeric vector indicating the number of bootstrapping samples. If bs.N=1, then bootstrapping is not used and bs.replace is ignored.}

\item{bs.replace}{a character string indicating whether bootstrapping is done with replacement. Default is TRUE.}

\item{varilist}{a character vector indicating variable names for which you wish to compute standardized mean difference. List variable names as a vector like c("VariableA","VariableB")}

\item{corrmethod}{a character string indicating which correlation coefficient is to be computed. These include "Pearson" (default), "Spearman", "Polychoric", or "Polyserial". For tetrachoric use "Polychoric" and for biserial use "Polyserial". This relies on wCorr::weightedCorr}
}
\description{
This function matches exposed units and unexposed units by a pre-specified buffer distance (Euclidean distance)
}
\examples{
estimateATT()
}
