% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.caliper.R
\name{match.caliper}
\alias{match.caliper}
\title{match.caliper function}
\usage{
match.caliper(
  data,
  bexp,
  gpsname,
  caliper_bw,
  exp.status = 1,
  index.status = 1,
  replace2 = TRUE
)
}
\description{
This internal function is
}
\examples{
match.caliper()
}
