% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.model.cv.R
\name{xgb.model.cv}
\alias{xgb.model.cv}
\title{xgb.model.cv function}
\usage{
xgb.model.cv(cv.local.N, max_depth, eta, subsample)
}
\description{
This internal function explores optimal parameters in xgb.train using cross-validation
xgb.model.cv()
}
