% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.model.fit.cv.R
\name{xgb.model.fit.cv}
\alias{xgb.model.fit.cv}
\title{xgb.model.fit.cv function}
\usage{
xgb.model.fit.cv(
  data,
  cv.objective,
  cv.nround,
  cv.nfold,
  cv.min.burnin,
  cv.max_depth,
  cv.eta,
  cv.gamma,
  cv.nthread,
  cv.subsample,
  cv.eval_metric,
  cv.colsample_bytree,
  cv.min_child_weight,
  cv.lambda,
  cv.lambda_bias,
  cv.alpha,
  cv.scale_pos_weight,
  early_stopping_rounds,
  cv.local.N
)
}
\description{
This internal function explores optimal parameters in xgb.train using cross-validation
xgb.model.fit.cv()
}
